requirements.txt
numpy
pandas
pyautogui==0.9.53
pillow==9.4.0